requirements.txt
npyscreen
PyYAML>=5.2